refacto: move Add to its own module
