+++ cargo-green/src/main.rs
@@ -11,6 +11,7 @@ use logging::{ENV_LOG, ENV_LOG_PATH};
 use rustc_wrapper::ENV;
 use tokio::process::Command;
 
+mod add;
 mod base_image;
 mod cargo_green;
 mod checkouts;
